[air] Remove checkpoint user guide and update key concepts and docstring (#27455)
