+++ doc/source/_toc.yml
@@ -15,7 +15,6 @@ parts:
     - file: ray-air/user-guides
       sections:
         - file: ray-air/preprocessors
-        - file: ray-air/checkpoints
         - file: ray-air/check-ingest
         - file: ray-air/config-scaling
         - file: ray-air/tuner
