+++ doc/source/ray-air/doc_code/air_key_concepts.py
@@ -65,6 +65,37 @@ best_result = result_grid.get_best_result()
 print(best_result)
 # __air_tuner_end__
 
+# __air_checkpoints_start__
+checkpoint = result.checkpoint
+print(checkpoint)
+# Checkpoint(local_path=..../checkpoint_000005)
+
+tuned_checkpoint = result_grid.get_best_result().checkpoint
+print(tuned_checkpoint)
+# Checkpoint(local_path=..../checkpoint_000005)
+# __air_checkpoints_end__
+
+# __checkpoint_adhoc_start__
+from ray.train.tensorflow import TensorflowCheckpoint
+import tensorflow as tf
+
+# This can be a trained model.
+def build_model() -> tf.keras.Model:
+    model = tf.keras.Sequential(
+        [
+            tf.keras.layers.InputLayer(input_shape=(1,)),
+            tf.keras.layers.Dense(1),
+        ]
+    )
+    return model
+
+
+model = build_model()
+
+checkpoint = TensorflowCheckpoint.from_model(model)
+# __checkpoint_adhoc_end__
+
+
 # __air_batch_predictor_start__
 from ray.train.batch_predictor import BatchPredictor
 from ray.train.xgboost import XGBoostPredictor
